% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{pepBayesCalls}
\alias{pepBayesCalls}
\alias{pepBayesCalls,matrix,numeric-method}
\alias{pepBayesCalls,pepBayesFit,numeric-method}
\title{Call peptide binding from pepBayes output.}
\usage{
pepBayesCalls(x, fdr, ...)

\S4method{pepBayesCalls}{pepBayesFit,numeric}(x, fdr = 0.05)

\S4method{pepBayesCalls}{matrix,numeric}(x, fdr = 0.05)
}
\arguments{
\item{x}{A matrix or pepBayesFit object.}

\item{fdr}{Desired false discovery rate.}
}
\value{
A matrix of booleans with TRUE denoting a positive call, and FALSE
a negative call.
}
\description{
Use posterior probability of response to generate a
call threshold that controls the false discovery rate of pepBayes positivity
calls.
}
\details{
Probabilities are sorted in descending order. The largest integer K
is selected such that the mean of the first K sorted probability values exceeds
the value (1 - \code{fdr}). The probability threshold is chosen to be the K-th
sorted probability value. Entries with probability greater than or equal to the
threshold value are considered positive calls, else an entry is a negative call.
}
\seealso{
\code{\link{pepBayesEcm}}
\code{\link{pepBayesMcmc}}
\code{\link{pepBayesFit-methods}}
}

