% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{pepBayesFit-methods}
\alias{pepBayesFit-methods}
\title{pepBayesFit methods.}
\description{
Methods for interacting with \code{\link{pepBayesFit}} objects.
}
\section{Accessors}{

\describe{
 \item{\code{position(x)}}{Return the position of peptides.}
 \item{\code{peptide(x)}}{Return the peptide amino acid sequences.}
 \item{\code{parTrace(x)}}{Return trace of fitted parameter values across iterations
of fitting algorithm.}
 \item{\code{ppb(x)}}{Return the estimated posterior probability of binding for each subject
/peptide combination.}
}
}

\section{pepBayes Analysis}{

\describe{
 \item{\code{pepBayesCalls(x, fdr)}}{Return calls for a pepBayesFit at the desired
\code{fdr} value.}
 \item{\code{pepBayesMcmc(x, ...)}}{Run \code{\link{pepBayesMcmc}}, starting the chain
at the last position of the parameters in \code{x}.}
 \item{\code{pepBayesEcm(x, ...)}}{Run \code{\link{pepBayesEcm}}, starting the optimization
at the last position of the parameters in \code{x}.}
}
}
\seealso{
\code{\link{pepBayesEcm}}
\code{\link{pepBayesMcmc}}
\code{\link{pepBayesCalls}}
}

